separate out web-related part 2
