+++ daemon/app/GHCSpecter/Render/Web.hs
@@ -22,10 +22,10 @@ import GHCSpecter.Channel.Common.Types (DriverId (..))
 import GHCSpecter.Channel.Outbound.Types (SessionInfo (..))
 import GHCSpecter.Data.Assets (HasAssets (..))
 import GHCSpecter.Data.Map (forwardLookup, keyMapToList, lookupKey)
-import GHCSpecter.Render.Components.Console qualified as Console
-import GHCSpecter.Render.SourceView qualified as SourceView
+import GHCSpecter.Render.Web.Console qualified as Console
 import GHCSpecter.Render.Web.ModuleGraph qualified as ModuleGraph
 import GHCSpecter.Render.Web.Session qualified as Session
+import GHCSpecter.Render.Web.SourceView qualified as SourceView
 import GHCSpecter.Render.Web.Timing qualified as Timing
 import GHCSpecter.Render.Util (divClass)
 import GHCSpecter.Server.Types (
