+++ daemon/app/GHCSpecter/Render/Web/ConsoleItem.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module GHCSpecter.Render.Components.ConsoleItem (
+module GHCSpecter.Render.Web.ConsoleItem (
   render,
 ) where
 
@@ -9,8 +9,8 @@ import Concur.Replica (onClick, style)
 import Data.Text qualified as T
 import Data.Tree (drawTree)
 import GHCSpecter.Data.GHC.Core (toBind)
-import GHCSpecter.Render.Components.GHCCore (renderTopBind)
 import GHCSpecter.Render.Util (divClass)
+import GHCSpecter.Render.Web.GHCCore (renderTopBind)
 import GHCSpecter.Server.Types (ConsoleItem (..))
 import GHCSpecter.UI.ConcurReplica.DOM (
   button,
