+++ daemon/src/GHCSpecter/Render/Components/GraphView.hs
@@ -5,27 +5,14 @@
 module GHCSpecter.Render.Components.GraphView (
   compileModuleGraph,
   compileGraph,
-  renderModuleGraph,
-  renderGraph,
 ) where
 
-import Concur.Core (Widget)
-import Concur.Replica (
-  classList,
-  onClick,
-  onMouseEnter,
-  onMouseLeave,
-  width,
- )
-import Concur.Replica.SVG.Props qualified as SP
 import Control.Lens ((^.), _1)
-import Data.Either.Extra (fromEither)
 import Data.IntMap (IntMap)
 import Data.IntMap qualified as IM
 import Data.Map qualified as M
-import Data.Maybe (catMaybes, fromMaybe)
+import Data.Maybe (fromMaybe)
 import Data.Text (Text)
-import Data.Text qualified as T
 import Data.Tuple (swap)
 import GHCSpecter.Channel.Common.Types (type ModuleName)
 import GHCSpecter.GraphLayout.Types (
@@ -47,11 +34,6 @@ import GHCSpecter.Graphics.DSL (
   TextPosition (..),
   ViewPort (..),
  )
-import GHCSpecter.Render.ConcurReplicaSVG (renderPrimitive)
-import GHCSpecter.Render.Util (xmlns)
-import GHCSpecter.UI.ConcurReplica.DOM (div, text)
-import GHCSpecter.UI.ConcurReplica.SVG qualified as S
-import GHCSpecter.UI.ConcurReplica.Types (IHTML)
 import GHCSpecter.UI.Types.Event (ModuleGraphEvent (..))
 import Prelude hiding (div)
 
@@ -171,71 +153,3 @@ compileGraph cond grVisInfo =
    in [Rectangle (0, 0) canvasWidth canvasHeight Nothing Nothing Nothing Nothing] -- just dummy for now
         ++ fmap edge (grVisInfo ^. gviEdges)
         ++ concatMap node (grVisInfo ^. gviNodes)
-
-renderModuleGraph ::
-  -- | key = graph id
-  IntMap ModuleName ->
-  -- | For each module, assign a double type value in [0, 1],
-  -- which will be shown as bar below the module node.
-  (ModuleName -> Double) ->
-  -- | Graph layout information
-  GraphVisInfo ->
-  -- | (focused (clicked), hinted (hovered))
-  (Maybe Text, Maybe Text) ->
-  Widget IHTML ModuleGraphEvent
-renderModuleGraph
-  nameMap
-  valueFor
-  grVisInfo
-  (mfocused, mhinted) =
-    let Dim canvasWidth canvasHeight = grVisInfo ^. gviCanvasDim
-        handlers hitEvent =
-          catMaybes
-            [ fmap (\ev -> ev <$ onMouseEnter) (hitEventHoverOn hitEvent)
-            , fmap (\ev -> ev <$ onMouseLeave) (hitEventHoverOff hitEvent)
-            , fmap (\ev -> fromEither ev <$ onClick) (hitEventClick hitEvent)
-            ]
-        scene = compileModuleGraph nameMap valueFor grVisInfo (mfocused, mhinted)
-        rexp = sceneElements scene
-        svgProps =
-          [ width (T.pack (show (canvasWidth + 100)))
-          , SP.viewBox
-              ( "0 0 "
-                  <> T.pack (show (canvasWidth + 100))
-                  <> " "
-                  <> T.pack (show (canvasHeight + 100))
-              )
-          , SP.version "1.1"
-          , xmlns
-          ]
-        svgElement =
-          S.svg
-            svgProps
-            ( S.style [] [text ".small { font: 6px Courier,monospace; } text { user-select: none; }"]
-                : fmap (renderPrimitive handlers) rexp
-            )
-     in div [classList [("is-fullwidth", True)]] [svgElement]
-
--- | render graph more simply
-renderGraph :: (Text -> Bool) -> GraphVisInfo -> Widget IHTML a
-renderGraph cond grVisInfo =
-  let Dim canvasWidth canvasHeight = grVisInfo ^. gviCanvasDim
-      rexp = compileGraph cond grVisInfo
-      svgProps =
-        [ width (T.pack (show (canvasWidth + 100)))
-        , SP.viewBox
-            ( "0 0 "
-                <> T.pack (show (canvasWidth + 100))
-                <> " "
-                <> T.pack (show (canvasHeight + 100))
-            )
-        , SP.version "1.1"
-        , xmlns
-        ]
-      svgElement =
-        S.svg
-          svgProps
-          ( S.style [] [text ".small { font: 6px Courier,monospace; } text { user-select: none; }"]
-              : fmap (renderPrimitive (\_ -> [])) rexp
-          )
-   in div [classList [("is-fullwidth", True)]] [svgElement]
